requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
redis==5.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.10.3
pydantic[email]
pydantic-settings==2.7.0
python-multipart==0.0.20
httpx==0.27.2
google-generativeai==0.8.3
authlib==1.3.0
itsdangerous==2.1.2
argon2-cffi==23.1.0
google-auth==2.29.0
